+++ src/constants/atoms/todayCategory.ts
@@ -1,0 +1,30 @@
+export const categoryList = [
+  {
+    name: '음식',
+    value: 'FOOD',
+  },
+  {
+    name: '음악',
+    value: 'MUSIC',
+  },
+  {
+    name: '관계',
+    value: 'RELATIONSHIP',
+  },
+  {
+    name: '음료',
+    value: 'BEVERAGE',
+  },
+  {
+    name: '문화생활',
+    value: 'CULTURE_LIFE',
+  },
+  {
+    name: '여행',
+    value: 'TRIP',
+  },
+  {
+    name: '나',
+    value: 'ME',
+  },
+]
